More docs, more everything

+++ examples/pppos/cell.h
@@ -7,14 +7,14 @@
 // specifications which standartises AT commands
 
 struct cell {
-  int state;                      // Current state, CELL_* enumeration
-  char *buf;                      // Received UART data buffer
-  int size;                       // Received UART buffer size
-  int len;                        // Received UART data len
-  const char **cmds;              // Setup commands, NULL terminated
-  int cmd;                        // Current setup command
-  unsigned long expire;           // Command expiration timestamp
-  void (*tx)(const void *, int);  // Transmit function
+  int state;                         // Current state, CELL_* enumeration
+  char *buf;                         // Received UART data buffer
+  size_t size;                       // Received UART buffer size
+  size_t len;                        // Received UART data len
+  const char **cmds;                 // Setup commands, NULL terminated
+  int cmd;                           // Current setup command
+  unsigned long expire;              // Command expiration timestamp
+  void (*tx)(const void *, size_t);  // Transmit function
 };
 
 enum { CELL_START, CELL_WAIT, CELL_SETUP, CELL_PPP };  // Modem states
